Move custom type hints to `types.py`
