+++ src/quacc/recipes/dftb/_base.py
@@ -15,8 +15,7 @@ if TYPE_CHECKING:
 
     from ase.atoms import Atoms
 
-    from quacc.schemas._aliases.ase import RunSchema
-    from quacc.types import Filenames, SourceDirectory
+    from quacc.types import Filenames, RunSchema, SourceDirectory
 
 LOG_FILE = "dftb.out"
 GEOM_FILE = "geo_end.gen"
