+++ src/quacc/recipes/emt/phonons.py
@@ -14,7 +14,7 @@ if TYPE_CHECKING:
 
     from ase.atoms import Atoms
 
-    from quacc.schemas._aliases.phonons import PhononSchema
+    from quacc.types import PhononSchema
 
 
 @flow
