+++ src/quacc/recipes/emt/slabs.py
@@ -14,7 +14,7 @@ if TYPE_CHECKING:
 
     from ase.atoms import Atoms
 
-    from quacc.schemas._aliases.ase import OptSchema, RunSchema
+    from quacc.types import OptSchema, RunSchema
 
 
 @flow
