+++ src/quacc/recipes/espresso/core.py
@@ -14,9 +14,7 @@ from quacc.recipes.espresso._base import run_and_summarize, run_and_summarize_op
 if TYPE_CHECKING:
     from ase.atoms import Atoms
 
-    from quacc.runners.ase import OptParams
-    from quacc.schemas._aliases.ase import RunSchema
-    from quacc.types import Filenames, SourceDirectory
+    from quacc.types import Filenames, OptParams, RunSchema, SourceDirectory
 
 BASE_SET_METAL = {
     "input_data": {
