+++ src/quacc/recipes/espresso/phonons.py
@@ -27,8 +27,12 @@ if TYPE_CHECKING:
 
     from ase.atoms import Atoms
 
-    from quacc.schemas._aliases.ase import RunSchema
-    from quacc.types import EspressoPhononDosSchema, Filenames, SourceDirectory
+    from quacc.types import (
+        EspressoPhononDosSchema,
+        Filenames,
+        RunSchema,
+        SourceDirectory,
+    )
