+++ src/quacc/recipes/gulp/_base.py
@@ -18,8 +18,8 @@ if TYPE_CHECKING:
 
     from ase.atoms import Atoms
 
-    from quacc.schemas._aliases.ase import RunSchema
-    from quacc.types import Filenames, SourceDirectory
+    from quacc.types import Filenames, RunSchema, SourceDirectory
+
 logger = logging.getLogger(__name__)
 
 GEOM_FILE_PBC = "gulp.cif"
