+++ src/quacc/recipes/gulp/core.py
@@ -10,8 +10,7 @@ from quacc.recipes.gulp._base import run_and_summarize
 if TYPE_CHECKING:
     from ase.atoms import Atoms
 
-    from quacc.schemas._aliases.ase import RunSchema
-    from quacc.types import Filenames, SourceDirectory
+    from quacc.types import Filenames, RunSchema, SourceDirectory
 
 
 @job
