+++ src/quacc/recipes/mlp/core.py
@@ -15,8 +15,7 @@ if TYPE_CHECKING:
 
     from ase.atoms import Atoms
 
-    from quacc.runners.ase import OptParams
-    from quacc.schemas._aliases.ase import OptSchema, RunSchema
+    from quacc.types import OptParams, OptSchema, RunSchema
 
 
 @job
