+++ src/quacc/recipes/mlp/phonons.py
@@ -20,7 +20,7 @@ if TYPE_CHECKING:
 
     from ase.atoms import Atoms
 
-    from quacc.schemas._aliases.phonons import PhononSchema
+    from quacc.types import PhononSchema
 
 
 @flow
