+++ src/quacc/recipes/onetep/_base.py
@@ -16,9 +16,7 @@ if TYPE_CHECKING:
 
     from ase.atoms import Atoms
 
-    from quacc.runners.ase import OptParams
-    from quacc.schemas._aliases.ase import RunSchema
-    from quacc.types import Filenames, SourceDirectory
+    from quacc.types import Filenames, OptParams, RunSchema, SourceDirectory
 
 
 def run_and_summarize(
