+++ src/quacc/recipes/orca/_base.py
@@ -17,9 +17,13 @@ if TYPE_CHECKING:
 
     from ase.atoms import Atoms
 
-    from quacc.runners.ase import OptParams
-    from quacc.schemas._aliases.cclib import cclibASEOptSchema, cclibSchema
-    from quacc.types import Filenames, SourceDirectory
+    from quacc.types import (
+        Filenames,
+        OptParams,
+        SourceDirectory,
+        cclibASEOptSchema,
+        cclibSchema,
+    )
 
 _LABEL = OrcaTemplate()._label  # skipcq: PYL-W0212
 LOG_FILE = f"{_LABEL}.out"
