+++ src/quacc/recipes/psi4/core.py
@@ -15,8 +15,7 @@ has_psi4 = bool(find_spec("psi4"))
 if TYPE_CHECKING:
     from ase.atoms import Atoms
 
-    from quacc.schemas._aliases.ase import RunSchema
-    from quacc.types import Filenames, SourceDirectory
+    from quacc.types import Filenames, RunSchema, SourceDirectory
 
 
 @job
