+++ src/quacc/recipes/vasp/_base.py
@@ -14,9 +14,13 @@ if TYPE_CHECKING:
 
     from ase.atoms import Atoms
 
-    from quacc.runners.ase import OptParams
-    from quacc.schemas._aliases.vasp import VaspASEOptSchema, VaspSchema
-    from quacc.types import Filenames, SourceDirectory
+    from quacc.types import (
+        Filenames,
+        OptParams,
+        SourceDirectory,
+        VaspASEOptSchema,
+        VaspSchema,
+    )
 
 
 def run_and_summarize(
