+++ src/quacc/recipes/vasp/mp.py
@@ -35,12 +35,12 @@ if TYPE_CHECKING:
 
     from ase.atoms import Atoms
 
-    from quacc.schemas._aliases.vasp import (
+    from quacc.types import (
         MPGGARelaxFlowSchema,
         MPMetaGGARelaxFlowSchema,
+        SourceDirectory,
         VaspSchema,
     )
-    from quacc.utils.files import SourceDirectory
 
 _MP_SETTINGS = {"VASP_INCAR_COPILOT": "off", "VASP_USE_CUSTODIAN": True}
