+++ src/quacc/recipes/vasp/qmof.py
@@ -18,9 +18,13 @@ from quacc.recipes.vasp._base import run_and_summarize, run_and_summarize_opt
 if TYPE_CHECKING:
     from ase.atoms import Atoms
 
-    from quacc.schemas._aliases.ase import OptSchema
-    from quacc.schemas._aliases.vasp import QMOFRelaxSchema, VaspSchema
-    from quacc.types import Filenames, SourceDirectory
+    from quacc.types import (
+        Filenames,
+        OptSchema,
+        QMOFRelaxSchema,
+        SourceDirectory,
+        VaspSchema,
+    )
 
 LOGGER = logging.getLogger(__name__)
