+++ src/quacc/runners/ase.py
@@ -37,58 +37,18 @@ has_sella = bool(find_spec("sella"))
 
 if TYPE_CHECKING:
     from pathlib import Path
-    from typing import Any, TypedDict
+    from typing import Any
 
     from ase.atoms import Atoms
     from ase.calculators.calculator import Calculator
     from ase.optimize.optimize import Dynamics
-    from numpy.random import Generator
 
-    from quacc.types import Filenames, SourceDirectory
-
-    class OptParams(TypedDict, total=False):
-        """
-        Type hint for `opt_params` used throughout quacc.
-        """
-
-        relax_cell: bool
-        fmax: float | None
-        max_steps: int
-        optimizer: Dynamics
-        optimizer_kwargs: dict[str, Any] | None
-        store_intermediate_results: bool
-        fn_hook: Callable | None
-        run_kwargs: dict[str, Any] | None
-
-    class MDParams(TypedDict, total=False):
-        """
-        Type hint for `md_params` used throughout quacc.
-        """
-
-        dynamics: MolecularDynamics
-        dynamics_kwargs: dict[str, Any] | None
-        steps: int
-        maxwell_boltzmann_kwargs: MaxwellBoltzmanDistributionKwargs | None
-        set_com_stationary: bool
-        set_zero_rotation: bool
-
-    class VibKwargs(TypedDict, total=False):
-        """
-        Type hint for `vib_kwargs` in [quacc.runners.ase.Runner.run_vib][].
-        """
-
-        indices: list[int] | None
-        delta: float
-        nfree: int
-
-    class MaxwellBoltzmanDistributionKwargs(TypedDict, total=False):
-        """
-        Type hint for `maxwell_boltzmann_kwargs` in [quacc.runners.ase.Runner.run_md][].
-        """
-
-        temperature_K: float
-        force_temp: bool
-        rng: Generator | None
+    from quacc.types import (
+        Filenames,
+        MaxwellBoltzmanDistributionKwargs,
+        SourceDirectory,
+        VibKwargs,
+    )
 
 
 class Runner(BaseRunner):
